requirements.txt
# Database Connection
psycopg2-binary==2.9.9

# Environment & Config
python-dotenv==1.0.0
PyYAML==6.0.1

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Data Analysis
pandas==2.1.3
numpy==1.26.2

# AI & NLP
sentence-transformers==2.2.2
torch>=2.2.0

# Dashboard
streamlit==1.28.2